Add scrollSquare component
