+++ src/components/layouts/ScrollSquare/ScrollSquare.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+const ScrollSquare = ({ clickScroll }) => {
+  return (
+    <div className="scrollSquareWrapper">
+      <a href="#first" className="fas fa-arrow-alt-circle-up">
+        {" "}
+      </a>
+    </div>
+  );
+};
+
+export default ScrollSquare;
