+++ src/components/layouts/Skills/Skills.js
@@ -14,11 +14,11 @@ const Skills = () => {
         <h2 className="skills__title">Skills</h2>
         <h3 className="skills__front">Front-End</h3>
         <p className="skills__text">
-          HTML, CSS, SASS, SCSS, BEM, RWD, ES6, JavaScript, React, ReactHooks,
-          ReactRouter, Redux
+          HTML, CSS, SASS, SCSS, BEM, RWD, ES6, JavaScript, TypeScript, React,
+          ReactHooks, ReactRouter, Redux, Webpack
         </p>
         <h3 className="skills__front">Others</h3>
-        <p className="skills__text">NPM, GIT, GitHub, GSAP, VSCode</p>
+        <p className="skills__text">NPM, YARN, GIT, GitHub, GSAP, VSCode</p>
       </div>
     </section>
   );
